[All] Fixed multiple issues during build on Linux
- Added some fixes introduced by RodoMa92 in PR198

- Lack of AVX2 extension (should be done differently in the future)
- Disable deprecated-volatile warning
- Added missing override in posix EventInfo, ImGui notification class and XContent class
- Removed not used XAudio2.h include in XMP
- Fixed missing switch-case in XObject
- Added fugly template in native_list.h
- Fixed multiple smaller issues

+++ src/xenia/gpu/trace_viewer.cc
@@ -899,9 +899,11 @@ void TraceViewer::DrawVertexFetcher(Shader* shader,
         } break;
         case xenos::VertexFormat::k_16_16_FLOAT: {
           auto e0 = LOADEL(uint32_t, 0);
-          ImGui::Text("%.2f", half_float::detail::uint16((e0 >> 16) & 0xFFFF));
+          ImGui::Text("%.2f", static_cast<float>(half_float::detail::uint16(
+                                  (e0 >> 16) & 0xFFFF)));
           ImGui::NextColumn();
-          ImGui::Text("%.2f", half_float::detail::uint16((e0 >> 0) & 0xFFFF));
+          ImGui::Text("%.2f", static_cast<float>(half_float::detail::uint16(
+                                  (e0 >> 0) & 0xFFFF)));
           ImGui::NextColumn();
         } break;
         case xenos::VertexFormat::k_32_32:
@@ -972,13 +974,17 @@ void TraceViewer::DrawVertexFetcher(Shader* shader,
         case xenos::VertexFormat::k_16_16_16_16_FLOAT: {
           auto e0 = LOADEL(uint32_t, 0);
           auto e1 = LOADEL(uint32_t, 1);
-          ImGui::Text("%.2f", half_float::detail::uint16((e0 >> 16) & 0xFFFF));
+          ImGui::Text("%.2f", static_cast<float>(half_float::detail::uint16(
+                                  (e0 >> 16) & 0xFFFF)));
           ImGui::NextColumn();
-          ImGui::Text("%.2f", half_float::detail::uint16((e0 >> 0) & 0xFFFF));
+          ImGui::Text("%.2f", static_cast<float>(half_float::detail::uint16(
+                                  (e0 >> 0) & 0xFFFF)));
           ImGui::NextColumn();
-          ImGui::Text("%.2f", half_float::detail::uint16((e1 >> 16) & 0xFFFF));
+          ImGui::Text("%.2f", static_cast<float>(half_float::detail::uint16(
+                                  (e1 >> 16) & 0xFFFF)));
           ImGui::NextColumn();
-          ImGui::Text("%.2f", half_float::detail::uint16((e1 >> 0) & 0xFFFF));
+          ImGui::Text("%.2f", static_cast<float>(half_float::detail::uint16(
+                                  (e1 >> 0) & 0xFFFF)));
           ImGui::NextColumn();
         } break;
         case xenos::VertexFormat::k_32_32_32_32_FLOAT:
